+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking
 
+- Fix a11y issue in teaser heading default body @sneridagh
+
 ### Feature
 
 - Be able to reset the block from the Grid block UI @sneridagh
@@ -11,6 +13,8 @@
 ### Bugfix
 
 ### Internal
+
+- Add new `headline` CSS class in teaser `head_title` @sneridagh
 
 ## 5.2.0 (2022-11-21)